refactor budget model and apply changes to tie a budget to a user

+++ src/app/modules/budgets/controller/budget.controller.ts
@@ -2,12 +2,15 @@ import { BudgetRepository } from '../repository/budget.repository';
 import { FastifyReply, FastifyRequest } from 'fastify';
 import { BudgetDto } from '../domain/dto/BudgetDto';
 import { BudgetMapper } from '../domain/mapper/Budget.mapper';
+import { UserService } from '../../users/services/user.service';
 
 export class BudgetController {
   private budgetRepository: BudgetRepository;
+  private userService: UserService;
 
-  constructor(repository: BudgetRepository) {
-    this.budgetRepository = repository;
+  constructor() {
+    this.budgetRepository = new BudgetRepository();
+    this.userService = new UserService();
   }
 
   async findAllBudgets(
@@ -15,15 +18,15 @@ export class BudgetController {
     reply: FastifyReply
   ): Promise<BudgetDto[]> {
     try {
-      const reqUserId = request.params['userId'] as string;
+      const userId = request.params['userId'] as string;
 
-      const budgetEntitys = await this.budgetRepository.findAll(reqUserId);
-      const budgetDtos = budgetEntitys.map((entity) =>
-        BudgetMapper.toDto(entity)
-      );
-
-      reply.statusCode = 202;
-      return budgetDtos;
+      if (await this.userService.userExists(userId)) {
+        const entities = await this.budgetRepository.findAll(userId);
+        reply.statusCode = 202;
+        return BudgetMapper.toDtos(entities);
+      } else {
+        reply.notFound();
+      }
     } catch (error: unknown) {
       if (error instanceof Error) {
         reply.internalServerError(error.message);
@@ -39,13 +42,16 @@ export class BudgetController {
     const userId = request.params['userId'];
 
     try {
-      const result = await this.budgetRepository.findById(id, userId);
-
-      if (result) {
-        reply.statusCode = 202;
-        return BudgetMapper.toDto(result);
+      if (await this.userService.userExists(userId)) {
+        const result = await this.budgetRepository.findById(id, userId);
+        if (result) {
+          reply.statusCode = 202;
+          return BudgetMapper.toDto(result);
+        } else {
+          reply.notFound();
+        }
       } else {
-        reply.notFound();
+        reply.conflict();
       }
     } catch (error: unknown) {
       if (error instanceof Error) {
@@ -62,13 +68,15 @@ export class BudgetController {
       const currentMonth = new Date(request.headers.date);
       const userId = request.params['userId'];
 
-      const results = await this.budgetRepository.findByMonth(
-        currentMonth,
-        userId
-      );
-      const dtos = results.map((result) => BudgetMapper.toDto(result));
-
-      return dtos;
+      if (await this.userService.userExists(userId)) {
+        const results = await this.budgetRepository.findByMonth(
+          currentMonth,
+          userId
+        );
+        return BudgetMapper.toDtos(results);
+      } else {
+        reply.notFound();
+      }
     } catch (error: unknown) {
       if (error instanceof Error) {
         reply.internalServerError();
@@ -77,14 +85,17 @@ export class BudgetController {
   }
 
   async createBudget(request: FastifyRequest, reply: FastifyReply) {
-    const reqBudget = request.body as BudgetDto;
-    const userId = request.params['userId'];
-
     try {
-      reply.statusCode = 201;
+      const reqBudget = request.body as BudgetDto;
+      const userId = request.params['userId'];
 
-      const entity = BudgetMapper.toEntity(reqBudget, userId);
-      await this.budgetRepository.create(entity);
+      if (await this.userService.userExists(userId)) {
+        const entity = BudgetMapper.toEntity(reqBudget, userId);
+        await this.budgetRepository.create(entity);
+        reply.statusCode = 201;
+      } else {
+        reply.conflict();
+      }
     } catch (error: unknown) {
       if (error instanceof Error) {
         reply.internalServerError();
@@ -93,18 +104,22 @@ export class BudgetController {
   }
 
   async updateBudget(request: FastifyRequest, reply: FastifyReply) {
-    const id = Number(request.params['id']);
-    const reqBudget = request.body as BudgetDto;
-    const userId = request.params['userId'];
-
-    if (id !== reqBudget.id) {
-      return reply.badRequest();
-    }
-
     try {
-      const entity = BudgetMapper.toEntity(reqBudget, userId);
-      await this.budgetRepository.update(entity);
-      reply.statusCode = 202;
+      const id = Number(request.params['id']);
+      const reqBudget = request.body as BudgetDto;
+      const userId = request.params['userId'];
+
+      if (id !== reqBudget.id) {
+        return reply.badRequest();
+      }
+
+      if (await this.userService.userExists(userId)) {
+        const entity = BudgetMapper.toEntity(reqBudget, userId);
+        await this.budgetRepository.update(entity);
+        reply.statusCode = 202;
+      } else {
+        reply.conflict();
+      }
     } catch (error: unknown) {
       if (error instanceof Error) {
         reply.internalServerError();
@@ -113,18 +128,22 @@ export class BudgetController {
   }
 
   async deleteBudget(request: FastifyRequest, reply: FastifyReply) {
-    const id = Number(request.params['id']);
-    const reqBudget = request.body as BudgetDto;
-    const userId = request.params['userId'];
-
-    if (id !== reqBudget.id) {
-      return reply.badRequest();
-    }
-
     try {
-      const entity = BudgetMapper.toEntity(reqBudget, userId);
-      await this.budgetRepository.delete(entity);
-      reply.statusCode = 202;
+      const id = Number(request.params['id']);
+      const reqBudget = request.body as BudgetDto;
+      const userId = request.params['userId'];
+
+      if (id !== reqBudget.id) {
+        return reply.badRequest();
+      }
+
+      if (await this.userService.userExists(userId)) {
+        const entity = BudgetMapper.toEntity(reqBudget, userId);
+        await this.budgetRepository.delete(entity);
+        reply.statusCode = 202;
+      } else {
+        reply.conflict();
+      }
     } catch (error: unknown) {
       if (error instanceof Error) {
         reply.internalServerError();
